MSF-395 - Stop Iptables if flag set on.

+++ disable_iptables_for_avahi_mdns.sh
@@ -1,16 +1,24 @@
 #!/bin/sh
 
+stop_ip_tables="$1"
+
 if systemctl is-active --quiet avahi-daemon; then
 
   echo "Avahi Daemon is running"
   if systemctl is-active --quiet iptables; then
 
-    if systemctl stop iptables && systemctl disable iptables; then
+    if [ "$stop_ip_tables" = "true" ]; then
 
-      echo "Iptables service is disabled for Avahi mDNS"
+      if systemctl stop iptables && systemctl disable iptables; then
+
+        echo "Iptables service is disabled for Avahi mDNS"
+      fi
+    else
+
+      echo "Iptables service Avahi mDNS will not be stopped"
     fi
   else
 
-    echo "Iptables service is already disabled for Avahi mDNS"
+    echo "Iptables service is stopped for Avahi mDNS"
   fi
 fi
